Apply the new Accordion comp and clean up the styles
